Prevent objects from falling through floor & add rebounding
Objects have a restituion property which determines the loss in speed with each collision.
Class navigation has been improved - points belonging to an obloid will be
associated with that obloid by reference. This means that higher level properties no
longer need to be passed down to methods for point calculations.